+++ lib/appium_lib_core/common/base/bridge/mjsonwp.rb
@@ -31,6 +31,7 @@ module Appium
           include Device::ScreenRecord::Command
           include Device::Device
           include Device::TouchActions
+          include Device::ExecuteDriver
 
           def commands(command)
             ::Appium::Core::Commands::MJSONWP::COMMANDS[command]
@@ -52,6 +53,13 @@ module Appium
 
           def send_actions(_data)
             raise Error::UnsupportedOperationError, '#send_actions has not been supported in MJSONWP'
+          end
+
+          # For Appium
+          # @param [Hash] id The id which can get as a response from server
+          # @return [::Selenium::WebDriver::Element]
+          def convert_to_element(id)
+            ::Selenium::WebDriver::Element.new self, element_id_from(id)
           end
         end # class MJSONWP
       end # class Bridge