+++ lib/appium_lib_core/common/base/bridge/w3c.rb
@@ -31,6 +31,7 @@ module Appium
           include Device::ScreenRecord::Command
           include Device::Device
           include Device::TouchActions
+          include Device::ExecuteDriver
 
           def commands(command)
             ::Appium::Core::Commands::W3C::COMMANDS[command]
@@ -131,6 +132,13 @@ module Appium
                   end
 
             ids.map { |id| ::Selenium::WebDriver::Element.new self, element_id_from(id) }
+          end
+
+          # For Appium
+          # @param [Hash] id The id which can get as a response from server
+          # @return [::Selenium::WebDriver::Element]
+          def convert_to_element(id)
+            ::Selenium::WebDriver::Element.new self, element_id_from(id)
           end
 
           # For Appium