+++ lib/appium_lib_core/common/command/common.rb
@@ -64,7 +64,8 @@ module Appium
         stop_recording_screen:      [:post, 'session/:session_id/appium/stop_recording_screen'],
         start_recording_screen:     [:post, 'session/:session_id/appium/start_recording_screen'],
         compare_images:             [:post, 'session/:session_id/appium/compare_images'],
-        is_keyboard_shown:          [:get,  'session/:session_id/appium/device/is_keyboard_shown']
+        is_keyboard_shown:          [:get,  'session/:session_id/appium/device/is_keyboard_shown'],
+        execute_driver:             [:post, 'session/:session_id/appium/execute_driver']
       }.freeze
 
       COMMAND_ANDROID = {